add abstract menu superclass
